refactor: Drop current implementation
The current implementation is problematic in multiple ways:

* Too many duplicate/non-generic content which is a maintenance and
  integration nightmare (like translation symbolic links)
* Contains Implementation in multiple languages, which is in most cases
  causing trouble instead of being appreciated
* Tools are not platform independent, while cooperating with one with a
  Windows operating system is a common use case

as a result, we should drop it entirely and start from scratch in order to
fix the problems.  Any previous implementations that are specific to a field
in particular should be migrated and maintained to a new project instead of
this one while we focus this one as a clean and simplified "common base".

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>