fix(deploy-dev-env): Work around Git's "detected dubious ownership..." error
The git repository might not have the same ownership user ID as with the running user in the development environment.  Ensure Git operations still work in such cases.

Note that this will create files with incorrect ownership settings in the Git repository, which is expected to be fixed by the users afterward.

Signed-off-by: 林博仁(Buo-ren Lin) <buo.ren.lin@gmail.com>